+++ include/Tools/Math/math_tools.h
@@ -3,15 +3,15 @@
 
 // Constants
 
-const float infinity = std::numeric_limits<float>::infinity();
-const float pi = 3.1415926535897932385f;
-const float inv_pi = 1/pi;
+const double infinity = std::numeric_limits<double>::infinity();
+const double pi = 3.1415926535897932385f;
+const double inv_pi = 1/pi;
 
-inline float degrees_to_radians(float degrees) {
+inline double degrees_to_radians(double degrees) {
 	return degrees * pi / 180.0;
 }
 
-inline float clamp(float x, float min, float max) {
+inline double clamp(double x, double min, double max) {
 	if (x < min) return min;
 	if (x > max) return max;
 	return x;
@@ -23,4 +23,4 @@ inline void idx_to_ij(int idx, int& i, int& j, int width)
 	j = idx / width;
 }
 
-inline float Lerp(float t, float v1, float v2) { return (1 - t) * v1 + t * v2; }
+inline double Lerp(double t, double v1, double v2) { return (1 - t) * v1 + t * v2; }
