+++ include/Tools/Math/vec3.h
@@ -16,7 +16,7 @@ class vec3 :public vec<3>
 public:
 	vec3() : vec<3>() {}
 	vec3(vec<3>& in_vec) :vec<3>(in_vec) {}
-	vec3(float e0, float e1, float e2)
+	vec3(double e0, double e1, double e2)
 	{
 		data[0] = e0; data[1] = e1; data[2] = e2;
 	}
