+++ include/Tools/Spectrum/RGBSpectrum.h
@@ -7,17 +7,17 @@
 class RGBSpectrum :public CoefficientSpectrum<3>
 {
 public:
-	RGBSpectrum(float v = 0.f) :CoefficientSpectrum<3>(v) {}
+	RGBSpectrum(double v = 0.f) :CoefficientSpectrum<3>(v) {}
 	RGBSpectrum(const CoefficientSpectrum<3>& v) :CoefficientSpectrum<3>(v) {}
 
 	static RGBSpectrum FromRGB(const vec3& rgb, SpectrumType type = SpectrumType::Illuminant);
 	static RGBSpectrum FromXYZ(const vec3& xyz, SpectrumType type = SpectrumType::Illuminant);
-	static RGBSpectrum FromSampled(const float* lambda, const float* v, int n);
+	static RGBSpectrum FromSampled(const double* lambda, const double* v, int n);
 
-	static float InterpolateSpectrumSamples(const float* lambda, const float* p, int i, const float x);
+	static double InterpolateSpectrumSamples(const double* lambda, const double* p, int i, const double x);
 
 	void ToRGB(vec3& rgb) const;
-	float y() const;
+	double y() const;
 
 	const RGBSpectrum& ToRGBSpectrum() const;
 
@@ -41,7 +41,7 @@ inline RGBSpectrum RGBSpectrum::FromXYZ(const vec3& xyz, SpectrumType type)
 	return FromRGB(rgb, type);
 }
 
-inline float RGBSpectrum::InterpolateSpectrumSamples(const float* lambda, const float* vals, int n, const float l)
+inline double RGBSpectrum::InterpolateSpectrumSamples(const double* lambda, const double* vals, int n, const double l)
 {
 	if (l <= lambda[0]) return vals[0];
 	if (l >= lambda[n - 1]) return vals[n - 1];
@@ -49,28 +49,28 @@ inline float RGBSpectrum::InterpolateSpectrumSamples(const float* lambda, const 
 	{
 		if (l >= lambda[i] && l <= lambda[i + 1])
 		{
-			float t = (l - lambda[i]) / (lambda[i + 1] - lambda[i]);
+			double t = (l - lambda[i]) / (lambda[i + 1] - lambda[i]);
 			return Lerp(t, vals[i], vals[i + 1]);
 		}
 	}
 }
 
-inline RGBSpectrum RGBSpectrum::FromSampled(const float* lambda, const float* v, int n)
+inline RGBSpectrum RGBSpectrum::FromSampled(const double* lambda, const double* v, int n)
 {
 	// Sort samples if unordered, use sorted for returned spectrum
 	if (!SpectrumSamplesSorted(lambda, v, n))
 	{
-		std::vector<float> slambda(&lambda[0], &lambda[n]);
-		std::vector<float> sv(&v[0], &v[n]);
+		std::vector<double> slambda(&lambda[0], &lambda[n]);
+		std::vector<double> sv(&v[0], &v[n]);
 		SortSpectrumSamples(&slambda[0], &sv[0], n);
 		return FromSampled(&slambda[0], &sv[0], n);
 	}
 	vec3 xyz = { 0, 0, 0 };
-	float yint = 0.f;
+	double yint = 0.f;
 	for (int i = 0; i < nCIESamples; ++i)
 	{
 		yint += CIE_Y[i];
-		float val = InterpolateSpectrumSamples(lambda, v, n, CIE_lambda[i]);
+		double val = InterpolateSpectrumSamples(lambda, v, n, CIE_lambda[i]);
 		xyz[0] += val * CIE_X[i];
 		xyz[1] += val * CIE_Y[i];
 		xyz[2] += val * CIE_Z[i];
@@ -88,7 +88,7 @@ inline void RGBSpectrum::ToRGB(vec3& rgb) const
 	rgb[2] = c[2];
 }
 
-inline float RGBSpectrum::y() const
+inline double RGBSpectrum::y() const
 {
 	vec3 xyz;
 	vec3 rgb;
